refactor: relocate admin import page
